requirements.txt
decorator==5.1.1
motor==3.0.0
netaddr==0.8.0
pydantic==1.10.2
pymongo==4.2.0
PyYAML==6.0
typing_extensions==4.4.0
validators==0.20.0
tldextract==3.4.0
pytz==2022.6
ray[default]==2.5.0
retry==0.9.2
GitPython==3.1.31 
slack-sdk==3.21.3
rocketry==2.5.1
boto3==1.28.20